+++ packages/react/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @openformat/react
 
+## 0.3.0-dev-39bb3fa
+
+### Patch Changes
+
+- [#96](https://github.com/open-format/js/pull/96) [`de229bb`](https://github.com/open-format/js/commit/de229bb53844c0aa91f3ae4e0824acaeed9dc595) Thanks [@tinypell3ts](https://github.com/tinypell3ts)! - set autoConnect to true in wagmi client
+
+- Updated dependencies [[`de229bb`](https://github.com/open-format/js/commit/de229bb53844c0aa91f3ae4e0824acaeed9dc595)]:
+  - @openformat/sdk@0.3.0-dev-39bb3fa
+
+## 0.3.0-dev-094b3d3
+
+### Patch Changes
+
+- [#94](https://github.com/open-format/js/pull/94) [`3bca926`](https://github.com/open-format/js/commit/3bca92659a99048bf4ee5351e8099b95d8b3b9d0) Thanks [@tinypell3ts](https://github.com/tinypell3ts)! - Handle default network in React Hook package provider
+
+- Updated dependencies [[`3bca926`](https://github.com/open-format/js/commit/3bca92659a99048bf4ee5351e8099b95d8b3b9d0)]:
+  - @openformat/sdk@0.3.0-dev-094b3d3
+
 ## 0.3.0-dev-854232e
 
 ### Minor Changes